+++ source/Nuke.Common/Git/GitRepositoryAttribute.cs
@@ -18,11 +18,16 @@ namespace Nuke.Common.Git
         public static GitRepository Value { get; private set; }
 
         [CanBeNull]
+        public string Branch { get; set; }
+
+        public string Remote { get; set; } = "origin";
+
+        [CanBeNull]
         public override object GetStaticValue ()
         {
             return Value = Value
                            ?? ControlFlow.SuppressErrors(() =>
-                               GitRepository.FromLocalDirectory(NukeBuild.Instance.RootDirectory));
+                               GitRepository.FromLocalDirectory(NukeBuild.Instance.RootDirectory, Branch, Remote.NotNull()));
         }
     }
 }